Extract P2P logic from TCP.Server
What?
=====

We extract the business logic out of `TCP.Server` and into the `P2P`
module.

Ideally, we should make the `P2P` module so that it need not have side
effects (sending packets and data via `TCP`). We can then rename the
`Adapter.TCP.Server` to `P2P.Server` and that will actually do the
sending of messages.