Move Adapter.TCP to P2P.Server
We combine ExWire.Adapter.TCP and ExWire.Adapter.TCP.Server into
ExWire.P2P.Server.

Why?
====

With all the core logic of peer to peer communication now in
`ExWire.P2P`, we simply want a p2p server to handle the state and to
subscribe to the tcp connection. The fact that the P2P server uses tcp
is an implementation detail, and should not be the primary driver of the
implementation. The separation of concerns is more clear now.

We update all places that use `Adapter.TCP` module and use the
`P2P.Server` instead.